seperate thrift module into 3 parts; extract Node Serialization from Server module into Cluster module
