+++ pom.xml
@@ -82,6 +82,8 @@
         <module>tsfile</module>
         <module>antlr</module>
         <module>thrift</module>
+        <module>thrift-cluster</module>
+        <module>thrift-sync</module>
         <module>service-rpc</module>
         <module>jdbc</module>
         <module>session</module>
