+++ thrift-sync/src/main/thrift/sync.thrift
